+++ example/lib/main.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:multi_screen_layout/multi_screen_layout.dart';
+import 'package:multi_screen_layout_example/android_device_posture_info_page.dart';
 import 'package:multi_screen_layout_example/master_detail_layout_example.dart';
 import 'package:multi_screen_layout_example/surface_duo_info_page.dart';
 import 'package:multi_screen_layout_example/two_page_layout_example.dart';
@@ -55,6 +56,16 @@ class MainPage extends StatelessWidget {
                   },
                 ),
                 ListTile(
+                  title: Text('Android Device Posture Info'),
+                  trailing: Icon(Icons.chevron_right),
+                  onTap: () {
+                    Navigator.push(
+                        context,
+                        MaterialPageRoute(
+                            builder: (_) => AndroidDevicePostureInfoPage()));
+                  },
+                ),
+                ListTile(
                   title: Text('Surface Duo Device Info'),
                   trailing: Icon(Icons.chevron_right),
                   onTap: () {
@@ -94,7 +105,12 @@ class SecondPage extends StatelessWidget {
         child: Column(
           mainAxisAlignment: MainAxisAlignment.spaceEvenly,
           children: <Widget>[
-            Text('This page only shows when spanned on 2 screens!'),
+            Text(
+              'Hello from page 2! This only displays when spanned across 2 '
+              'displays or the device posture is half opened. Also known as '
+              'Samsung\'s Flex Mode.',
+              textAlign: TextAlign.center,
+            ),
           ],
         ),
       ),