+++ lib/src/platform_handler.dart
@@ -1,15 +1,26 @@
+import 'dart:async';
 import 'dart:convert';
 import 'dart:io';
 
 import 'package:flutter/services.dart';
+import 'package:multi_screen_layout/src/devices/android_standard.dart';
 import 'package:multi_screen_layout/src/devices/surface_duo.dart';
 import 'package:multi_screen_layout/src/models.dart';
+import 'package:multi_screen_layout/src/util.dart';
 
-const MethodChannel _channel = const MethodChannel('multi_screen_layout');
+const _methodChannel = const MethodChannel('multi_screen_layout');
+const _devicePostureEventChannel =
+    const EventChannel('multi_screen_layout_device_posture');
 
 /// Provides access to the platform to get multi screen information
 class MultiScreenPlatformHandler {
   MultiScreenPlatformHandler._();
+
+  /// Directly listen to changes in device posture
+  static Stream<DevicePosture> onDevicePostureChanged =
+      _devicePostureEventChannel
+          .receiveBroadcastStream()
+          .map((value) => devicePostureFromInt((value as int)));
 
   /// Provides access to SurfaceDuo specific information
   static SurfaceDuoPlatformHandler surfaceDuo = SurfaceDuoPlatformHandler._();
@@ -17,13 +28,12 @@ class MultiScreenPlatformHandler {
   /// Gets all relevant multi screen information in one call to the platform
   static Future<MultiScreenLayoutInfoModel> getInfoModel() async {
     if (!Platform.isAndroid) return null;
-    final result = await _channel.invokeMethod<String>('getInfoModel');
+    final result = await _methodChannel.invokeMethod<String>('getInfoModel');
     if (result == null) {
       return null;
     } else {
-      return MultiScreenLayoutInfoModel(
-        surfaceDuoInfoModel: SurfaceDuoInfoModel.fromJson(jsonDecode(result)),
-      );
+      var infoModel = PlatformInfoModel.fromJson(jsonDecode(result));
+      return MultiScreenLayoutInfoModel.fromPlatform(infoModel);
     }
   }
 }
@@ -34,26 +44,29 @@ class SurfaceDuoPlatformHandler {
 
   static Future<bool> getIsDual() async {
     if (!Platform.isAndroid) return false;
-    final isDual = await _channel.invokeMethod<bool>('isDualScreenDevice');
+    final isDual =
+        await _methodChannel.invokeMethod<bool>('isDualScreenDevice');
     return isDual;
   }
 
   static Future<bool> getIsSpanned() async {
     if (!Platform.isAndroid) return false;
-    final isAppSpanned = await _channel.invokeMethod<bool>('isAppSpanned');
+    final isAppSpanned =
+        await _methodChannel.invokeMethod<bool>('isAppSpanned');
     return isAppSpanned;
   }
 
   static Future<double> getHingeAngle() async {
     if (!Platform.isAndroid) return null;
-    final hingeAngle = await _channel.invokeMethod<double>('getHingeAngle');
+    final hingeAngle =
+        await _methodChannel.invokeMethod<double>('getHingeAngle');
     return hingeAngle;
   }
 
   static Future<NonFunctionalBounds> getNonFunctionalBounds() async {
     if (!Platform.isAndroid) return null;
     final result =
-        await _channel.invokeMethod<String>('getNonFunctionalBounds');
+        await _methodChannel.invokeMethod<String>('getNonFunctionalBounds');
     if (result == null) {
       return null;
     } else {