+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.0
+
+* Added support for Samsung Flex Mode for both `TwoPageLayout` and `MasterDetailLayout`
+* For both `TwoPageLayout` and `MasterDetailLayout` you can now disable their behavior on specific types of devices
+
 ## 0.0.3
 
 * Added `MasterDetailLayout`
@@ -8,4 +13,4 @@
 
 ## 0.0.1
 
-* First release with TwoPageLayout and Surface Duo support
+* First release with `TwoPageLayout` and Surface Duo support